DO NOT MERGE ANYWHERE: Account for emergency alert notifications
Since we are using the addToNotifications path for displaying
the alert for both emergency and non emergency the emergency
sound gets overridden by the non emergency sound played from
addToNotifications if the config is enabled.

This patch accounts for the message being an emergency message
because if it is then the alert sound will have already
been dispatched to the CellBroadcastAlertAudio service.

Bug: 37192772
Test: Manual with CellBroadcastTests test app
Change-Id: I02cdfb46eca281e04fb0e2a2b15fc1cb86b05579