Split out certain extreme CMAS warnings for KO.
Regulatory requirements in Korean dictate that the title of the CMAS
dialog must be different depending on the severity, urgency, and
certainty as well as on the class of the CMAS message. To satisfy
these requirements, two new string resources are added. I'll work
with the localization team to ensure these are translated suitably
for the Korean locale.

Bug: 64950858
Change-Id: I12036504c3308455c1eeb83017ba4ad7edf7aef8
(cherry picked from commit 1b98ead3313491c7eea6243b8ec3b388a826e7a7)