Make notification not swipeable
Make the emergency notification alert not swipeable. Also allow
users to see the content of the entire alert messages from the
notification.

Note the notification is only shown when users not explicitly dismiss
the alert message (e.g. tap home button to navigate away from the
alert message). When the user tap the notification, the original
alert message diaglog pops up again and the notificaiton is removed.

Test: manual
Bug: 67925992
Change-Id: I3cd6d82778e5b31be65d1439905bc1312f172150