Fixed that emergency alert does not vibrate in DnD mode
The emergency alert does not vibrate in Do-Not-Disturb mode
even when "Use full volume" option is turned on in emergency
alert menu.

Now it vibrates when "Use full volume" option is turned on
and the "viration" option is turned on in emergency alert
menu. Also play in alarm channel which most people allow when
DnD mode is on.

Test: Manual
Bug: 117590597
Change-Id: Ic54a8307580380ca89db69a58653665c3c19a88f