Fixed that extra vibration was played
The alert vibration did not stop at 10.5 seconds because
the pre-recorded 10.5 seconds tone took more than 10.5s to
finish playing. It caused the vibration pattern to repeat
and vibrated for extra 0.2 seconds.

The solution was that we only play the vibration once if
we also play the tone once, which is the case for most
of the countries.

Test: Manual
Bug: 118358949
Change-Id: I76ee30e75b36f78b4e9b1fc16ee422bb139f1677