Fixed using full volume default value not changed
The using full volume default settings was written
when boot up first time. The settings was based on the
SIM inserted at that time and will not change even
after SIM swapped.

Fixed by adding another flag to indicate if the settings
has been changed by the user or not. If the user hasn't
changed the settings, always use the default value from
the resource overlay.

Test: Manual
Change-Id: I8ca16446e161d7b5a7a4bcc3ec30680f2a6fd5df