New alert is hidden by WEA opt-out dialog
New coming emergency alert is displayed under the opt-out
dialog so it can not be read by user.

This is due to opt-out dialog display at the top layer of
the screen, so the new coming alert is displayed under the
opt-out.
As a solution, hide the opt-out dialog when new alert coming
and show it again after all alerts read.

Bug: 141768151

Change-Id: I95c7d9f4b4e0e813b9d781372b0478bac11598c3
Merged-in: I95c7d9f4b4e0e813b9d781372b0478bac11598c3