Added isPreferenceVisible support
Instead of removing the preference from the screen, use
the new tag isPreferenceVisible to control whether the
toggle should be displayed or not. The toggles will be
displayed only if the following are conditions are all met.
1. isPreferenceVisible is true.
2. Associated cell broadcast channels are configured.

Not all toggles support isPreferenceVisible in this CL.
They can be added if needed in the future.

Test: Manual
Bug: 148644685
Change-Id: I94ebe578697f7ea21e2efd49eecac26544313acc