Do not require BROADCAST_SMS
This cannot be granted to CellBroadcastService, which sends out the
broadcast, since we are removing the telephony permission role. But on
the bright side, it also is not needed since the broadcast is protected.

Since we no longer require the permission (the broadcasts which we don't
want spoofed are protected broadcasts) we remove the
PrivilegedCellBroadcastReceiver.

Bug: 148482594
Test: verify protected broadcast cannot be sent by non-system
Change-Id: I02af3fae702eb0af725caacf28f5a7e324ea468c
(cherry picked from commit 1903ad8a744578d1a326829f526f2a53217481ee)