Show disaster icon for ETWS in Japan
This commit is a handoff to Jack who is taking over the bug.
Show disaster icons for ETWS depending on type of disaster.
Note: Icons are not googley but are instead standard images
used within Japan.  ETWS is only used in Japan and so the
decision to use these specific icons made sense.

Bug: 130401071
Test: Manual + unit tests
Change-Id: Ie3ad6b136405b1f6d2af50e1f7233db7a81b3401