Add @InstrumentationTest annotation
CellBroadcastContentProviderTest requires the actual content provider,
which means we cannot include it in the OEM test apk, as it's not a unit
test. This commit adds the @InstrumentationTest so that it will be
excluded from that apk.

Bug: 160208930
Test: atest CellBroadcastContentProviderTest
Change-Id: I544f753247016ffec4ecfdea08dce75878794511
(cherry picked from commit c71dbd42bd934ebe0e1e621c7bce9c32f72d6416)