Chile requirement adjustment
1. channel 919 should have the emergency sound rather than sms
notification sound.
2. For test channel 519 and 4380, The message must be similar
to the ones received from short messaging system, and should
not follow the display, sound and vibration standard established
by the use of the channel CL-Alerta Local.
3. Fix the issue that changeStreamVolume for notification will
fail due to not granted Do Not Disturb Access. We don't need
notification stream type anymore as we have a local copy of
notification sound. Thus it should be treated as other alarm type.

Bug: 161511367
Test: Manual test
Change-Id: Ic8a7810e66cee42b2bf830fd35f26037533691c0