At config to show year in date
If the config to show the date in the dialog is true, we also check this
config to determine whether to show the year as part of the date. This
is set to true for Peru and Chile.

Bug: 163003175
Test: manually verify date shows
Change-Id: I856e001feb3d81a54a617312a33b308d868adae0
(cherry picked from commit 188e3a342a5e43a9c90b37dc8a853bdc429eefaf)