Reset CBSettings preferences on carrier change
We only process carrier changes for the default subscription. If the
default subscription changes, we won't update settings until next
carrier config broadcast for default sub (usually next boot).

If the CellBroadcastSettings activity is open or in recent-apps when the
reset occurs, the values will not reset until the activity is closed and
re-opened.

Bug: 166073620
Bug: 159680560
Bug: 162700005
Test: manual
Change-Id: I4af5ce73082412b32053e6c125479ce98149db53