Always mark db transaction as successful for data migration.
It is called from onCreate() which creates the main table, and
onCreate() is withing a begin/end transcation block. Any nested
transcation failing reverts the other transactions in the block
as well and as a result the table will not be created if migration
transaction is not marked as successful.

Test: manual
Bug: 166379976
Change-Id: I9bde2d9f6eac4803d8eb5423e99bc7ced5f7dc14