Revert "Add overlayable ids for OEM customization"
This reverts commit 439e24e514c793819516a57d55ae25669dfd2451.

This CL is only a partial revert -- it removes the IDs and overlayable attrs
which are part of the androidx migration, not the other IDs and attrs.

Note that on master the CellBroadcastListViewTheme and
CellBroadcastListItemTheme do not exist, so we do not put them in
overlayable.

Reason for revert: stability and not needed by OP on short term, see b/163879473

Change-Id: I55ab6b7f5dbe3367dd34b76a1007239732fdcf64