disable secondary channels for Peru
channel 50(together with channel 919) belong to the secondary
configuration and per spec "The devices that support the channels
indicated in the main configuration do not require support the channels
indicated in the secondary configuration. This will also resolve the
issue that CMAS alert incorrectly triggered in roaming
(Peru SIM + Brazilian network). Because channel 50 was enabled when
inserting Peru SIM card but channel 50 is reserved for other purpose
(area info update) in and are broadcasted constantly in Brazil.

Bug: 171531247
Test: Manual
Change-Id: I0ef45fc04b355289306013a0e834955022f5f583