Fixed alert tone not playing when screen is off
When screen is off and wake up with screen lock,
the dialog's onStop() is called immediately. This caused
alert tone stop playing immediately. Before introducing the
more complex logic, removing the logic that stops
tone playing when diaglog pushed to background. This
will cause the side effect that after alert is shown, if
the user press home button while the alert tone is still
playing, the alert tone will be played until the end. We
think this is acceptable.

Fix: 171611796
Test: Manual
Change-Id: I09048c4ca296e4af62965442c115d2c1ffa7685a