fixed dialog title language for Saudi and UAE (temp solution)
Saudi and UAE mandate to show fixed dialog title language: Arabia
for the primary channels and English for the secondary channel.
The right solution should be read language code from the network
message. However the testlab in those regions didn't setup the language
code correctly thus fails the TA testing. we are hardcoding the language
code as a temp workaround to unblock OEMs. We should remove this once
test scripts was updated.

Bug: 167366175
Test: Manual
Change-Id: Ieed7451f960c02e16d7d4d13279a01fe9f24d79e