Red exclamation point icon in dialog disappears onPause
Red exclamation point icon in dialog disappears for a while
when moving to the link. The issue here is inside onPause,
the imageView for the warning icon is set to invisible, not
only the red exclamation point icon disappeared but text shifted to
left. There is no need to make imageview disappear inside onPause
given the dialog is no longer in the foreground.

Bug: 172787290
Test: Manual test
Change-Id: Id6d351f0aabdcdabe79aa338d3398eb67cd70371