make the from_dialog notification clickable
in some rare scenarios, alert activity was never resumed and user
only get this non-clickable from_dialog notification. As a result,
users didn't see the pop-up dialog and the notification is non-clickable
so that users can't read the complete messsage. to fix this, we treat
the "from_dialog" notification same as the notification posted when
users press the home key. tap the notification will open the alert
dialog,with following special handlings:
1. tap the notification while alert dialog is in the foreground, no
duplicate mesages added to the dialog.
2. tap the notification does not dismiss this notification. the only
ways to dismiss this notification is either user dismiss the dialog or
user swipe away the notification

Bug: 172619487
Test: Manual
Change-Id: Ic10d12d58b5f06b498c64bc9e8b2df3fdca6cfce