Revert Saudi Arabia language override
Saudi Arabia had a temporary mLanguageCode override in order to fix an
issue with testers not setting the correct language code. This was reverted
when our understanding was that the correct language code was now being
set, however we are revertting again (only for SA this time, not UAE)
because it seems the test still needs this workaround.

Reason for revert: partial revert for b/179624700
Reverted Changes:
Idf6b19732:[automerge] Revert "fixed dialog title language fo...
Ic82323bde:Revert "fixed dialog title language for Saudi and ...

Test: manual
Change-Id: I3daa23892747f4f19d721704758113342bddb4f5