Compile CellBroadcastReceiver with module_current
This sdk_version is one created especially for modules, and includes
@SystemApi(client = MODULE_LIBRARIES) on top of normal @SystemApi.

Note that to call these APIs the app must be allowed to access hidden
APIs, which normal APKs are not but APK-in-apexes are.

This allows removing framework-annotations from the module sources,
which avoids follow-on problems of depending on shared util libraries.

Bug: 198418216
Test: m com.android.cellbroadcast
Merged-In: I2be92e9a42e104987519bfd3c617c443d9bfece1
Change-Id: I2be92e9a42e104987519bfd3c617c443d9bfece1