Reset the cache of CellBroadcastChannelRange
The cache of CellBroadcastChannelRange should be reset when sim
removed or loaded. The intent of ACTION_SIM_CARD_STATE_CHANGED is
not good for such case as SIM_STATE_PRESENT may be sent when sim
ready but record is not loaded, and no more broadcast when loaded.
ACTION_CARRIER_CONFIG_CHANGED can cover all the cases.

Bug: 213027609
Test: atest CellBroadcastReceiverUnitTests:CellBroadcastReceiverTest
Test: manually to remove/insert sim
Change-Id: I8029376c474c635cc4e374a87e6e496e71ff867a