Enable local test channels when preference is not shown but default is True
According to Qatar new reuqirment, default preference of local test channels was changed to be Enabled
and the preferene is not shown to user.
But mcc mnc is not changed in upgrade case, so preference value was remained,
so the change of default preference value cannot be applied
, while it is OK in new device/factory reset case/SIM swap case.
So, we need workaround that enables local test channels when preference is not shown but default value is True.
And this same workaround is also added in the logic that checks if message should be displayed when message is received.

Bug: 204256732
Bug: 222606767
Test: atest com.android.cellbroadcastreceiver.unit.CellBroadcastConfigServiceTest#testEnablingLocalTestChannels
Test: atest com.android.cellbroadcastreceiver.unit.CellBroadcastAlertServiceTest#testShouldDisplayMessage
Test: Manual
Change-Id: I2e18796f097f397b1795f9012c5a43ccd37b4c87