Use listview API for getting multi-selection status
CBR crash occurred due to ArrayIndexOutOfBoundsException in CellBroadcastListActivity because there is mismatch between mSelectedMessages and actual selected state of listview when activity is re-created on multi-window change.
For resolving this mismatch, we can directly get selected states with listview API instead of CellBroadcastListActivity managing mSelectedMessages.
In this change, codes regarding mSelectedMessages are removed and listview API like getCheckedItemPositions() is used

Bug: 216739398
Bug: 231674936
Test: Manual
Test: atest com.android.cellbroadcastreceiver.unit.CellBroadcastListActivityTest
Change-Id: I7b7cf5cddf3110eb01b368c2078d2dfb3756b1c7