Bypass cellbroadcastcompliancetest when mockmodem is not allowed in userversion
In user version, mockmodem is not allowed at default. If user turns on "Allow Mockmodem" in developer options,
mockmodem can be used.
So, let's use assumeTrue instead of test failure so that test can be bypassed along with the guide about "Allow mockmodem" of developer options
when mockmodem is not allowed.

Bug: 240230154
Test: m mts && mts-tradefed run mts-cellbroadcast
Change-Id: I2858fa6a4073a8f9f83739d2e219d9ccc585de83