Add exported flag to exposed test runtime receiver
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U
and registering a receiver for non-system broadcasts must
specify either the exported or not exported flag when invoking
 #registerReceiver; if one of these flags is not provided, the
platform will throw a SecurityException. This commit updates
the receiver in CellBroadcastConfigTest with the RECEIVER_EXPORTED
flag to maintain the existing behavior for this test.

Bug: 234659204
Test: atest CellBroadcastReceiverComplianceTests
Change-Id: Ic45056462b0e653c9306cacf2d3003d8f8688b9c